+++ CHANGELOG.md
@@ -26,6 +26,8 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 ### Fixed
 
+* Fixed: ":!{cmd}" execute {cmd} with 'shell' error (Windows)
+* Fixed: ":!!" repeat last ":!{cmd}" error (Windows)
 * Fixed: plugin unloading error
 * Fixed [#119](https://github.com/NeoVintageous/NeoVintageous/issues/119): Loosing user settings when toggling ctrl keys